Split types module up into atoms and sequences submodules
